refactor(legacy): decompose cli.ts into src/cli/, delete legacy/
Split the 497-line god file into focused modules:
- cli/index.ts — entry point: parseArgs + subcommand dispatch + main()
- cli/helpers.ts — requireArg, withDaemonErrors, apiAction, printJsonOr
- cli/service.ts — macOS LaunchAgent management
- cli/cron.ts — cron CRUD commands
- cli/logs.ts — log viewer with --follow support
- cli/init.ts — config wizard

Cleanup:
- Remove empty "Sessions command" dead section
- Fix redundant dynamic import of fs/promises (already imported at top)
- Make printJsonOr a pure function (json flag as param, not closure)
- Use dynamic imports for subcommands to reduce startup cost
- Delete src/legacy/ directory entirely
